Handling roles on arbitrary objects in the rights editor
The rights editor now supports editing rights on arbitrary roles on any
object, not just queues and globally.  (It still requires you extend RT's
core roles with your own, of course.)

Adds a new method on RT::Groups, LimitToRolesForObject, which
limits by Domain and Instance appropriately for the object given.  The
new method abstracts away more of the details of roles on objects.