change how we deal with decoding structured fields
we use MIME::Field::ParamVal already to deal with
continued params (RFC2231). This module allows us
to deal with separate parameters, decode them and
deal with quotes inside or absence of those without
guessing.

There is a bug in the module [1] - wrong parsing
of quoted strings and wrong quoting on stringify.
It was fixed, patch is on rt.cpan.org waiting for
merge and release of the module.

Similar situation with From, Cc, Bcc fields. Use
Email::Address to parse and split, then decode
and format.