Merge core dependencies into one list, and eliminate duplicates
PSGI and MASON are core dependencies which cannot be disabled using
command-line flags; fold them into CORE.  Though CLI and MAILGATE are
similarly required, and cannot be disabled using command-line flags,
keep them separate for now, as they are standalone binaries which are
intended to be able to be distributed separately.  Keeping a separate
list of their dependencies is thus of use.