Disable selection of individual objects when "apply globally" is checked
Unapplied scrips and CFs show both an "apply globally" checkbox as well
as collection lists of available individual objects.  If "apply
globally" is checked, prevent checking individual objects as well
before the page is submitted.  The API already prevents both changes
from going through, so this is simply a bit of icing on the admin
interface.